feat(strategies): rename erc721BalanceOfIssuedBy to arianeeErc721BalanceOfIssuedBy
