+++ src/strategies/strategies.ts
@@ -2,7 +2,7 @@ export { strategy as erc20BalanceOf } from './erc-20-balance-of';
 export { strategy as erc721BalanceOf } from './erc-721-balance-of';
 export { strategy as erc721OwnerOf } from './erc-721-owner-of';
 export { strategy as erc721NotOwnerOf } from './erc-721-not-owner-of';
-export { strategy as erc721BalanceOfIssuedBy } from './erc-721-balance-of-issued-by';
+export { strategy as arianeeErc721BalanceOfIssuedBy } from './arianee-erc-721-balance-of-issued-by';
 export { strategy as isExactAddress } from './is-exact-address';
 export { strategy as roomOwner } from './room-owner';
 export { strategy as erc1155BalanceOfBatch } from './erc-1155-balance-of-batch';
